requirements.txt
pyyaml
terrascript
namegenerator
emails
jinja2
python-terraform
python-gnupg